+++ libs/csengine/polygon/polygon.cpp
@@ -1485,8 +1485,8 @@ bool csPolygon3D::DoPerspective (const csTransform& trans,
         rquad++;
       while (quad != rquad)
       {
-        epointx = (quad&2)           ?  MAX_VALUE : -MAX_VALUE;
-        epointy = (quad==0||quad==3) ?  MAX_VALUE : -MAX_VALUE;
+        epointx = (float) (quad&2)           ?  MAX_VALUE : -MAX_VALUE;
+        epointy = (float) (quad==0||quad==3) ?  MAX_VALUE : -MAX_VALUE;
         dest->AddVertex (epointx, epointy);
         quad = (quad+1)&3;
       }
@@ -1501,8 +1501,8 @@ bool csPolygon3D::DoPerspective (const csTransform& trans,
         rquad++;
       while (quad != rquad)
       {
-        epointx = (quad&2)           ? -MAX_VALUE :  MAX_VALUE;
-        epointy = (quad==0||quad==3) ?  MAX_VALUE : -MAX_VALUE;
+        epointx = (float) (quad&2) ? -MAX_VALUE :  MAX_VALUE;
+        epointy = (float) (quad==0||quad==3) ?  MAX_VALUE : -MAX_VALUE;
         dest->AddVertex (epointx, epointy);
         quad = (quad+1)&3;
       }
